+++ Editor/Attribute/MetaDataDrawer/InspectorButton/InspectorButtonAttributeDrawer.cs
@@ -1,39 +1,4 @@
-#if UNITY_EDITOR && !UNITY_6000_0_OR_NEWER
-using System.Linq;
-using System.Collections.Generic;
-using System.Reflection;
-using UnityEngine;
-using UnityEditor;
-using Jeomseon.Extensions;
-using Jeomseon.Editor;
-
-namespace Jeomseon.Attribute.Editor
-{
-    using Editor = UnityEditor.Editor;
-
-    internal sealed class InspectorButtonAttributeDrawer : IObjectEditorAttributeDrawer
-    {
-        private readonly List<KeyValuePair<string, MethodInfo>> _buttonMethods = new();
-
-        public void OnEnable(Editor editor)
-        {
-            _buttonMethods.AddRange(EditorReflectionHelper
-                .GetMethodsFromAttribute<InspectorButtonAttribute>(editor.target)
-                .Select(method =>
-                    new KeyValuePair<string, MethodInfo>(
-                        method.GetCustomAttribute<InspectorButtonAttribute>().ButtonName,
-                        method)));
-        }
-
-        public void OnInspectorGUI(Editor editor)
-        {
-            _buttonMethods
-                .Where(kvp => GUILayout.Button(kvp.Key))
-                .ForEach(kvp => kvp.Value.Invoke(editor.target, null));
-        }
-    }
-}
-#elif UNITY_EDITOR && UNITY_6000_0_OR_NEWER
+#if UNITY_EDITOR
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -46,9 +11,9 @@ namespace Jeomseon.Attribute.Editor
     using Editor = UnityEditor.Editor;
 
     /// <summary>
-    /// Unity 6000.x 이상에서 InspectorButtonAttribute를 처리하는 전역 드로어.
-    /// InspectorWindow 내부 구조를 건드리지 않고
-    /// Editor.finishedDefaultHeaderGUI 공식 API만 사용한다.
+    /// InspectorButtonAttribute를 처리하는 전역 드로어입니다.
+    /// Unity 내부 InspectorWindow 구현을 리플렉션하지 않고
+    /// 오래전부터 제공되는 Editor.finishedDefaultHeaderGUI 공식 API만 사용합니다.
     /// </summary>
     [InitializeOnLoad]
     internal static class InspectorButtonHeaderDrawer
@@ -59,7 +24,8 @@ namespace Jeomseon.Attribute.Editor
 
         static InspectorButtonHeaderDrawer()
         {
-            // 모든 인스펙터 헤더가 그려진 뒤 호출되는 이벤트에 구독
+            // TODO(UX): 헤더가 아닌 본문 하단 배치가 필요해지면 CustomEditor를 강제하지 말고
+            // 공식 Editor 확장 지점을 제공하는 별도 opt-in 베이스 Editor 방식을 검토해야 합니다.
             Editor.finishedDefaultHeaderGUI += OnFinishedDefaultHeaderGUI;
         }
 
@@ -159,4 +125,3 @@ namespace Jeomseon.Attribute.Editor
     }
 }
 #endif
-